refactor(bitbucket): split vendor modules by responsibility
